Separate Player and Enemy logic, add Health and Projectile
