+++ Assets/Characters/Enemy/EnemyCharacter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 17d2035cb202548489ae6e3cfd875b4a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
