+++ Assets/Characters/Player/PlayerCharacter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 60ed3a66c62871549871b0b739c0b4ba
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
